@dataset_stat/requirements.txt
numpy>=1.23
pandas>=2.0
biopython>=1.80
matplotlib>=3.7
seaborn>=0.13
tqdm>=4.66

